requirements.txt
tf-keras
scikit-learn
spacy
sentence-transformers
chromadb
faiss-cpu